+++ config/hard/Candies/code/CGCode.js
@@ -1,0 +1,8 @@
+var inputs = readline().split(' ');
+const n = parseInt(inputs[0]);
+const k = parseInt(inputs[1]);
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');