+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "X egg problem".
 - configuration for "Paper-folding curve".
 - configuration for "SHA-256 Hash".
+- configuration for "Candies".
 
 ## [1.12.0] - 2025-02-28
 ### Added